pyproject.toml
[project]
name = "nosql-puc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.3.0",
    "psycopg2-binary>=2.9.10",
    "pymongo>=4.13.2",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.41",
]
